+++ CHANGELOG.rst
@@ -1,5 +1,7 @@
 Development
 ===========================
+* Test WSGI environment does not use asynchronous stream
+* Code cleanup
 
 
 Ver. 1.0.1 - 2015-Jun-03